requirements.txt
fastapi
uvicorn
ollama
sqlalchemy
pymysql
pandas
load_dotenv